POC/requirements.txt
torch==2.1.2
torchvision==0.16.2
opencv-python==4.9.0.80
scikit-image==0.22.0
timm==0.9.12
tqdm==4.66.1
pyyaml==6.0.1
wandb==0.16.2
numpy>=1.24.0
matplotlib>=3.7.0
pillow>=9.5.0
scipy>=1.10.0